requirements.txt
blake3
bcrypt
argon2-cffi